Apply file rename suggestions
