+++ arm9/source/patches/platform/compactflash-common/CompactFlashReadWriteSectorPatchCode.s
