+++ arm9/source/patches/platform/compactflash-common/CompactFlashStatusFunctionsPatchCode.h
