+++ arm9/source/patches/platform/m3cf/M3CFLoaderPlatform.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "common.h"
-#include "M3CFLockUnlockCard.h"
+#include "M3CFLockUnlockCardPatchCode.h"
 #include "../compactflash-common/CompactFlashCommonLoaderPlatform.h"
 
 /// @brief Implementation of LoaderPlatform for the DATEL line of flashcarts
