+++ arm9/source/patches/platform/m3cf/M3CFLockUnlockCardPatchCode.s
