+++ arm9/source/patches/platform/supercardcf/SuperCardCFLockUnlockCardPatchCode.s
